+++ forge/src/main/java/tocraft/walkers/forge/WalkersForge.java
@@ -1,6 +1,5 @@
 package tocraft.walkers.forge;
 
-import net.minecraftforge.common.MinecraftForge;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.loading.FMLEnvironment;
 import tocraft.walkers.Walkers;